Revert "Merge branch 'master' of https://github.com/ballerina-platform/ballerina-lang into http+io"
This reverts commit af76cdafe8798a7f6cfa6fd10dc7b7e214a80710, reversing
changes made to c5010bf8ee5c898aed52681ccfeed4e8c2961208.

+++ bvm/ballerina-runtime/src/main/java/org/ballerinalang/jvm/values/DecimalValue.java
@@ -159,7 +159,7 @@ public class DecimalValue {
                decimalValue.compareTo(BLangConstants.BINT_MIN_VALUE_BIG_DECIMAL_RANGE_MIN) > 0;
     }
 
-    public int byteValue() {
+    public long byteValue() {
         switch (valueKind) {
             case NOT_A_NUMBER:
                 throw new BallerinaException(BallerinaErrorReasons.NUMBER_CONVERSION_ERROR,
@@ -177,7 +177,7 @@ public class DecimalValue {
             throw new BallerinaException(BallerinaErrorReasons.NUMBER_CONVERSION_ERROR,
                                          "'decimal' value '" + value + "' cannot be converted to 'byte'");
         }
-        return (int) intVal;
+        return intVal;
     }
 
     private static boolean isByteLiteral(long longValue) {
